requirements.txt
openai>=1.0.0
python-pptx
python-docx
streamlit
python-dotenv
